requirements.txt
aiogram>=3.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
python-dotenv>=1.0.0

# Для webhook режима
aiohttp>=3.9.0

# Для тестов (опционально)
pytest>=7.0.0
pytest-asyncio>=0.21.0
